requirements.txt
python-telegram-bot==20.7
ccxt>=4.0.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0

